runtime/bebob: convert to binary crate
It is natural to use binary crate for runtime executables instead of
library crate.
