+++ runtime/bebob/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 edition = "2018"
-name = "bebob-runtime"
+name = "snd-bebob-ctl-service"
 version = "0.1.0"
 # For publishing.
 authors = ["Takashi Sakamoto"]
 description = """
-Implementation of protocol for ASIC in BridgeCo Enhanced Break Out Box and vendor-specific models
+Runtime implementation to control audio and music unit in IEEE 1394 bus which ALSA bebob driver
+(snd-bebob) supports.
 """
 license = "GPL-3.0-or-later"
 publish = false
