+++ runtime/bebob/src/main.rs
@@ -21,7 +21,8 @@ mod yamaha_terratec;
 use {
     alsa_ctl_tlv_codec::DbInterval,
     alsactl::{prelude::*, *},
-    core::{card_cntr::*, dispatcher::*, *},
+    clap::Parser,
+    core::{card_cntr::*, cmdline::*, dispatcher::*, LogLevel, *},
     firewire_bebob_protocols as protocols,
     glib::{source, Error, FileError},
     hinawa::{
@@ -46,7 +47,7 @@ enum Event {
     StreamLock(bool),
 }
 
-pub struct BebobRuntime {
+struct BebobRuntime {
     unit: (SndUnit, FwNode),
     model: BebobModel,
     card_cntr: CardCntr,
@@ -309,4 +310,27 @@ impl BebobRuntime {
             self.timer = None;
         }
     }
+}
+
+struct BebobServiceCmd;
+
+#[derive(Parser, Default)]
+#[clap(name = "snd-bebob-ctl-service")]
+struct Arguments {
+    /// The numeric identifier of sound card in Linux sound subsystem.
+    card_id: u32,
+
+    /// The level to debug runtime, disabled as a default.
+    #[clap(long, short, arg_enum)]
+    log_level: Option<LogLevel>,
+}
+
+impl ServiceCmd<Arguments, u32, BebobRuntime> for BebobServiceCmd {
+    fn params(args: &Arguments) -> (u32, Option<LogLevel>) {
+        (args.card_id, args.log_level)
+    }
+}
+
+fn main() {
+    BebobServiceCmd::run()
 }
